+++ canvas/color.c
@@ -6,7 +6,7 @@
 /*   By: abelayad <abelayad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 12:52:29 by abelayad          #+#    #+#             */
-/*   Updated: 2023/09/30 10:44:06 by abelayad         ###   ########.fr       */
+/*   Updated: 2023/12/03 22:24:43 by abelayad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,6 @@ const t_color	g_red = (t_color){1, 0, 0};
 const t_color	g_green = (t_color){0, 1, 0};
 const t_color	g_blue = (t_color){0, 0, 1};
 
-t_color	ft_color(double r, double g, double b)
-{
-	return ((t_color){r, g, b});
-}
-
 t_color	ft_add_colors(t_color a, t_color b)
 {
 	bool	test;
@@ -32,15 +27,6 @@ t_color	ft_add_colors(t_color a, t_color b)
 		a.r + b.r,
 		a.g + b.g,
 		a.b + b.b
-	});
-}
-
-t_color	ft_sub_colors(t_color a, t_color b)
-{
-	return ((t_color){
-		a.r - b.r,
-		a.g - b.g,
-		a.b - b.b
 	});
 }
