+++ include/camera.h
@@ -6,7 +6,7 @@
 /*   By: abelayad <abelayad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 13:57:19 by abelayad          #+#    #+#             */
-/*   Updated: 2023/10/31 15:41:36 by abelayad         ###   ########.fr       */
+/*   Updated: 2023/12/03 22:23:34 by abelayad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,11 +17,8 @@
 
 t_matrix	ft_view_transform_inverse(t_tuple from, t_tuple forward);
 t_camera	ft_camera(t_camera raw_camera);
-// t_ray		ft_ray_for_pixel(int x, int y, t_camera c);
-// void		ft_render(t_window window, t_world *w, t_camera c);
 void		ft_render(t_png_img *img, t_world *w, int phase);
 t_xnode		*ft_intersect_world(t_world *w, t_ray r);
-uint32_t	ft_merge_colors(int r, int g, int b, int a);
-void		ft_free_objs_and_tex(t_world *w);
+void		free_world(t_world *w);
 
 #endif
