--- include/ft_print.h
@@ -1,10 +1,0 @@
-#ifndef FT_PRINT_H
-# define FT_PRINT_H
-
-void	ft_print_tuple(t_tuple t);
-void	ft_print_color(t_color c);
-void	ft_print_ray(t_ray r);
-void	ft_print_xs(t_xnode *xs);
-
-
-#endif
