+++ include/lsts.h
@@ -6,14 +6,13 @@
 /*   By: abelayad <abelayad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 13:45:23 by abelayad          #+#    #+#             */
-/*   Updated: 2023/10/27 20:33:22 by abelayad         ###   ########.fr       */
+/*   Updated: 2023/12/04 09:52:59 by abelayad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef LSTS_H
 # define LSTS_H
 
-void		ft_print_xs(t_xnode *xnode);
 t_xnode		*ft_xnew(t_obj *o, double x);
 void		ft_xadd_back(t_xnode **xlst, t_xnode *n);
 t_xnode		*ft_sort_xs(t_xnode *head);
