+++ include/minirt.h
@@ -6,7 +6,7 @@
 /*   By: abelayad <abelayad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 13:59:57 by abelayad          #+#    #+#             */
-/*   Updated: 2023/11/22 22:02:54 by abelayad         ###   ########.fr       */
+/*   Updated: 2023/12/04 11:42:26 by abelayad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,8 @@
 # include <png.h>
 # include <ctype.h>
 # include <libft.h>
+# include <pthread.h>
+# include <png.h>
 # include "minirt_types.h"
 # include "tuples.h"
 # include "canvas.h"
@@ -28,11 +30,15 @@
 # include "lsts.h"
 # include "camera.h"
 # include "light_shading.h"
-# include "world.h"
 # include "ft_png.h"
 # include "ft_parse.h"
-# include "ft_printers.h"
-# include <pthread.h>
-// #include <sys/sysctl.h>
+
+typedef struct s_chunk
+{
+	t_png_img	*img;
+	t_world		*w;
+	int			num;
+	pthread_t	thread;
+}	t_chunk;
 
 #endif
