+++ include/ray_intersect.h
@@ -6,7 +6,7 @@
 /*   By: abelayad <abelayad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 13:56:37 by abelayad          #+#    #+#             */
-/*   Updated: 2023/10/27 20:27:52 by abelayad         ###   ########.fr       */
+/*   Updated: 2023/12/04 11:21:55 by abelayad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,6 @@ t_xnode	*ft_intersect_cylinder(t_obj *o, t_ray r);
 t_xnode	*ft_intersect_cone(t_obj *o, t_ray r);
 t_xnode	*ft_intersect_world(t_world *w, t_ray r);
 bool	ft_pt_bound_cycone(t_obj *o, t_ray r, double t);
-bool	ft_pt_bound_cone(t_obj *o, t_ray r, double t);
 t_xnode	*ft_hit(t_xnode *xlst);
 void	ft_intersect_cone_xs(t_xnode **xs, t_quadratics q, t_obj *o, t_ray r);
 void	ft_free_xs(t_xnode **xs_lst);
