+++ matrices/matrix_transformations.c
@@ -6,7 +6,7 @@
 /*   By: abelayad <abelayad@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/20 15:30:40 by abelayad          #+#    #+#             */
-/*   Updated: 2023/09/20 15:30:41 by abelayad         ###   ########.fr       */
+/*   Updated: 2023/12/04 10:57:51 by abelayad         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,42 +31,6 @@ t_matrix	ft_scale(double x, double y, double z)
 			{x, 0, 0, 0},
 			{0, y, 0, 0},
 			{0, 0, z, 0},
-			{0, 0, 0, 1}
-		}
-	});
-}
-
-t_matrix	ft_rotate_x(double rad)
-{
-	return ((t_matrix){
-		4, {
-			{1, 0, 0, 0},
-			{0, cos(rad), -sin(rad), 0},
-			{0, sin(rad), cos(rad), 0},
-			{0, 0, 0, 1}
-		}
-	});
-}
-
-t_matrix	ft_rotate_y(double rad)
-{
-	return ((t_matrix){
-		4, {
-			{cos(rad), 0, sin(rad), 0},
-			{0, 1, 0, 0},
-			{-sin(rad), 0, cos(rad), 0},
-			{0, 0, 0, 1}
-		}
-	});
-}
-
-t_matrix	ft_rotate_z(double rad)
-{
-	return ((t_matrix){
-		4, {
-			{cos(rad), -sin(rad), 0, 0},
-			{sin(rad), cos(rad), 0, 0},
-			{0, 0, 1, 0},
 			{0, 0, 0, 1}
 		}
 	});
