--- scenes/1.rt
@@ -1,5 +1,0 @@
-# A 0.2 255,255,255
-C -1,0,-3 0,0,0 180
-l 0,0,-3.1 255,255,255
-# l 5,-5,5 255,255,255
-pl 0,-1,0 0,1,0 255,255,255 .5 0 0 1
